Split form-components into text-input and membership-select
