+++ src/membership/components/NewMemberForm.jsx
@@ -12,8 +12,8 @@ import getMemberPrice from '../../lib/get-member-price'
 import { buyMembership, getPurchaseData } from '../../payments/actions'
 import StripeCheckout from '../../payments/components/stripe-checkout'
 import * as PaymentPropTypes from '../../payments/proptypes'
-import { MembershipSelect } from './form-components'
 import MemberForm from './MemberForm'
+import MembershipSelect from './membership-select'
 
 class NewMemberForm extends React.Component {
   static propTypes = {
